New: readstat_variable_add_missing_double_value
Define one or more values to be treated as missing using the new
readstat_variable_add_missing_double_value function. (E.g., a value
representing "No response".) The first three of these values will be
exported by the SAV writer. In the future it would be handy to support
ranges of values, but the encoding logic is a little more complicated.