SAS writer: Fix crash/hang issues (return error instead)
When a row of data wouldn't fit into the page size (currently defined to
be 4096 bytes, need to make this more flexible in the future), ReadStat
would either hang or crash with a divide-by-zero exception. This commit
adds checks so that READSTAT_ERROR_ROW_IS_TOO_WIDE_FOR_PAGE is returned
instead.