XPORT reader: Allow character encoding to be specified
The encoding of XPORT files is either UTF-8 or a Windows Latin variant,
depending on the source computer. XPORT files from SAS 9.2 and later
"stamp" the file with the encoding (TODO), but for earlier files,
it must be specified by the destination. Use the usual machinery,
`readstat_set_file_character_encoding` and friends, so that string
values are transliterated appropriately.

Fixes #172