Fix issues with truncated UTF-8 in SAV files
This was more work than I expected, but to avoid problems with truncated
UTF-8, store the name and longname as unconverted bytes for most of the
SAV file processing, and only convert them to the requested encoding
right before the variable handler is called.

Also fix a bug with unconverted labels in POR files.