Fix unicode non-ASCII name check with unsigned char
In case 'char' is unsigned (which happens on some architectures), then
checks like 'str[i] > 0' will always be true (for non-NULL characters,
of course). Since we are interested to check only for the values <= 127,
then change the check to cast the char to unsigned, so we can always
check that the value is lower than 0x80 (128).