Add strict prototypes for functions with no arguments. (#295)
Using compile flags `-Werror -Wstrict-prototypes` (with Clang at least),
the "prototypes" of some functions are flagged as an error. This is
because of the difference between:
```
	int no_args_please(void);
```
and
```
	int no_args_please();
```
Where the former is a strict prototype that says "no arguments are
allowed". With the compile flags mentioned, ReadStat fails to
compile with error messages like this:

```
In file included from src/readstat_bits.c:9:
src/readstat_bits.h:10:29: error: a function declaration without a prototype is deprecated in all versions of C [-Werror,-Wstrict-prototypes]
int machine_is_little_endian();
```

This PR adds the missing `void` in a handful of places where Clang
complains about it with those strict compile flags.